app/requirements.txt
psycopg2-binary==2.9.4
pandas==1.5.0
python-dotenv==0.21.0